float

+++ api/Models/EventLoc.cs
@@ -8,8 +8,8 @@ namespace api.Models
     {
         public int Id{get;set;}         //id
         public int EventId{get;set;}    //Event ID
-        public float Lat{get;set;}      //latitude
-        public float Long{get;set;}     //longitude
+        public double Lat{get;set;}      //latitude
+        public double Long{get;set;}     //longitude
         public String Nome{get;set;}    //Nome da localização
         public String Desc{get;set;}    //Descrição
         public Photo Img{get;set;}      //Foto da localização
